requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
asn1crypto==1.5.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
bip-utils==2.9.3
bitarray==3.0.0
cbor2==5.6.5
cdp-agentkit-core==0.0.11
cdp-langchain==0.0.13
cdp-sdk==0.15.0
certifi==2024.12.14
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.4.1
ckzg==2.0.1
coincurve==20.0.0
constantly==23.10.4
crcmod==1.7
cryptography==44.0.0
cytoolz==1.0.1
daphne==4.1.2
dataclasses-json==0.6.7
distro==1.9.0
Django==4.2.18
django-cors-headers==4.6.0
django-elasticsearch-dsl==8.0
django-environ==0.12.0
djangorestframework==3.15.2
ecdsa==0.19.0
ed25519-blake2b==1.4.1
elastic-transport==8.17.0
elasticsearch==8.17.1
elasticsearch-dsl==8.17.1
eth-account==0.13.4
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.1.0
eth-typing==5.1.0
eth-utils==5.2.0
eth_abi==5.2.0
factory_boy==3.3.1
Faker==35.0.0
frozenlist==1.5.0
gunicorn==21.2.0
h11==0.14.0
hexbytes==1.3.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.0.0
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.15
langchain-community==0.3.15
langchain-core==0.3.31
langchain-openai==0.2.14
langchain-text-splitters==0.3.5
langgraph==0.2.67
langgraph-checkpoint==2.0.10
langgraph-sdk==0.1.51
langsmith==0.3.1
marshmallow==3.26.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==2.2.2
oauthlib==3.2.2
openai==1.60.1
orjson==3.10.15
packaging==24.2
parsimonious==0.10.0
pluggy==1.5.0
propcache==0.2.1
py-sr25519-bindings==0.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
PyJWT==2.10.1
PyNaCl==1.5.0
pyOpenSSL==25.0.0
pytest==7.4.4
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyunormalize==16.0.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rlp==4.0.1
service-identity==24.2.0
setuptools==75.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
sqlparse==0.5.3
sse-starlette==2.2.1
standard-imghdr==3.13.0
starlette==0.45.3
tavily-python==0.5.0
tenacity==9.0.0
tiktoken==0.8.0
toolz==1.0.0
tqdm==4.67.1
tweepy==4.15.0
Twisted==24.11.0
twitter-langchain==0.0.11
txaio==23.1.1
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
web3==7.7.0
websockets==13.1
wheel==0.44.0
yarl==1.18.3
zope.interface==7.2
zstandard==0.23.0
